+++ inst/include/dplyr/Result/DelayedProcessor.h
@@ -3,119 +3,170 @@
 
 namespace dplyr{
 
-    template <typename CLASS, typename Data>
+    template <typename CLASS>
     class DelayedProcessor_Base {
        public:
-           typedef typename Data::group_iterator group_iterator ;
-
            DelayedProcessor_Base(){}
            virtual ~DelayedProcessor_Base(){}
 
-           virtual SEXP delayed_process( const Data& map, SEXP first_result, CLASS*, group_iterator git ) = 0;
+           virtual bool handled(int i, const RObject& chunk ) = 0 ;
+           virtual bool can_promote(const RObject& chunk ) = 0 ;
+           virtual DelayedProcessor_Base* promote(int i, const RObject& chunk) = 0 ;
+           virtual SEXP get() = 0;
     } ;
 
     template <int RTYPE>
-    typename Rcpp::traits::storage_type<RTYPE>::type strong_as( SEXP x ){
-        return as< typename Rcpp::traits::storage_type<RTYPE>::type >(x) ;
+    bool valid_conversion(int rtype){
+        return false ;    
     }
-
+    
     template <>
-    inline int strong_as<INTSXP>( SEXP x){
-        if( TYPEOF(x) == REALSXP ){
-            stop( "loss of precision when attempting to convert a %s to an integer", get_single_class(x) ) ;
+    inline bool valid_conversion<REALSXP>( int rtype ){
+        switch( rtype ){
+        case REALSXP:
+        case INTSXP:
+        case LGLSXP:
+            return true ;
+        default: break ;
         }
-        return as<int>(x) ;
+        return false ;
     }
-
-    template <>
-    inline int strong_as<LGLSXP>( SEXP x){
-        if( TYPEOF(x) == REALSXP || TYPEOF(x) == INTSXP ){
-            stop( "loss of precision when attempting to convert a %s to an logical", get_single_class(x) ) ;
+    
+    template <> 
+    inline bool valid_conversion<INTSXP>( int rtype ){
+        switch( rtype ){
+        case INTSXP:
+        case LGLSXP:
+            return true ;
+        default: break ;
         }
-        return as<int>(x) ;
+        return false ;
     }
-
-    template <int RTYPE, typename CLASS, typename Data>
-    class DelayedProcessor : public DelayedProcessor_Base<CLASS, Data> {
+    
+    template <int RTYPE>
+    inline bool valid_promotion(int rtype) {
+        return false ;    
+    }
+    
+    template <> 
+    inline bool valid_promotion<INTSXP>( int rtype ){
+        return rtype == REALSXP ;
+    }
+    
+    template <> 
+    inline bool valid_promotion<LGLSXP>( int rtype ){
+        return rtype == REALSXP || rtype == INTSXP ;
+    }
+    
+    template <int RTYPE, typename CLASS>
+    class DelayedProcessor : public DelayedProcessor_Base<CLASS> {
     public:
         typedef typename Rcpp::traits::storage_type<RTYPE>::type STORAGE ;
         typedef Vector<RTYPE> Vec ;
-        typedef typename Data::group_iterator group_iterator ;
-
-        DelayedProcessor(int first_non_na_) : first_non_na(first_non_na_){}
-
-        virtual SEXP delayed_process(const Data& gdf, SEXP first_result, CLASS* obj, group_iterator git) {
-
-            int n = gdf.ngroups() ;
-            Vector<RTYPE> res = no_init(n) ;
-            STORAGE* ptr = res.begin() ;
-
-            int i=0 ;
-            for( ; i<first_non_na; i++){
-                ptr[i] = Vec::get_na() ;
-            }
-            ptr[i] = strong_as<RTYPE>( first_result );
-            ++git ;
-            i++ ;
-            for( ; i<n; i++, ++git )
-                ptr[i] = strong_as<RTYPE>( obj->process_chunk(*git) ) ;
-            return res ;
+        
+        DelayedProcessor( int first_non_na, SEXP first_result, int ngroups_) :
+            res( no_init(ngroups_) ) 
+        {
+            std::fill( res.begin(), res.begin() + first_non_na, Vec::get_na() ); 
+            res[first_non_na] = as<STORAGE>( first_result ) ;
         }
-
+        
+        DelayedProcessor( int i, const RObject& chunk, SEXP res_ ) :
+            res( as<Vec>( res_ ) )
+        {
+            res[i] = as<STORAGE>(chunk) ;        
+        }
+        
+        virtual bool handled(int i, const RObject& chunk ) {
+            int rtype = TYPEOF(chunk) ;
+            if( valid_conversion<RTYPE>(rtype) ){
+                res[i] = as<STORAGE>( chunk ) ;
+                return true ;                                         
+            } else {
+                return false ;    
+            }
+        }
+        
+        virtual bool can_promote(const RObject& chunk ) {
+            return valid_promotion<RTYPE>( TYPEOF(chunk) ) ;    
+        }
+        virtual DelayedProcessor_Base<CLASS>* promote(int i, const RObject& chunk){
+            int rtype = TYPEOF(chunk) ;
+            switch( rtype ){
+            case INTSXP:  return new DelayedProcessor<INTSXP , CLASS>(i, chunk, res ) ;
+            case REALSXP: return new DelayedProcessor<REALSXP, CLASS>(i, chunk, res ) ;
+            case CPLXSXP: return new DelayedProcessor<CPLXSXP, CLASS>(i, chunk, res ) ;
+            default: break ;
+            }
+            return 0 ;
+        }
+        
+        virtual SEXP get() {
+            return res ;    
+        }
+        
+        
     private:
-        int first_non_na ;
+        Vec res ;
+        
+        
     } ;
 
-    template <typename CLASS, typename Data>
-    class DelayedProcessor<STRSXP, CLASS, Data> : public DelayedProcessor_Base<CLASS, Data> {
+    template <typename CLASS>
+    class DelayedProcessor<STRSXP, CLASS> : public DelayedProcessor_Base<CLASS> {
     public:
-        typedef typename Data::group_iterator group_iterator ;
+        DelayedProcessor(int first_non_na_, SEXP first_result, int ngroups) : 
+            res(ngroups)
+        {
+            res[first_non_na_] = as<String>(first_result) ;
+        }
 
-        DelayedProcessor(int first_non_na_) : first_non_na(first_non_na_){}
-
-        virtual SEXP delayed_process(const Data& gdf, SEXP first_result, CLASS* obj, group_iterator git) {
-            int n = gdf.ngroups() ;
-            CharacterVector res(n) ;
-            int i=0 ;
-            for( ; i<first_non_na; i++) res[i] = NA_STRING ;
-            res[i] = STRING_ELT(first_result, 0 ) ;
-            ++git ;
-            i++ ;
-            for( ; i<n; i++, ++git )
-                res[i] = STRING_ELT( obj->process_chunk(*git), 0) ;
-            return res ;
+        virtual bool handled(int i, const RObject& chunk ) {
+            res[i] = as<String>(chunk) ; 
+            return true ;
+        }                                 
+        virtual bool can_promote(const RObject& chunk ) {
+            return false ;    
+        }
+        virtual DelayedProcessor_Base<CLASS>* promote(int i, const RObject& chunk) {
+            return 0 ;    
+        }
+        virtual SEXP get() {
+            return res ;    
         }
 
     private:
-        int first_non_na ;
+        CharacterVector res ;
     } ;
 
-    template <typename CLASS, typename Data>
-    class DelayedProcessor<VECSXP, CLASS, Data> : public DelayedProcessor_Base<CLASS, Data> {
+    template <typename CLASS>
+    class DelayedProcessor<VECSXP, CLASS> : public DelayedProcessor_Base<CLASS> {
     public:
-        typedef typename Data::group_iterator group_iterator ;
+        DelayedProcessor(int first_non_na_, SEXP first_result, int ngroups) : 
+            res(ngroups)
+        {
+            res[first_non_na_] = maybe_copy(VECTOR_ELT(first_result, 0)) ;   
+        }
 
-        DelayedProcessor(int first_non_na_) : first_non_na(first_non_na_){}
-
-        virtual SEXP delayed_process(const Data& gdf, SEXP first_result, CLASS* obj, group_iterator git) {
-            int n = gdf.ngroups() ;
-            List res(n) ;
-            int i=0 ;
-            res[0] = maybe_copy(VECTOR_ELT(first_result, 0)) ;
-            ++git ;
-            i++ ;
-            for( ; i<n; i++, ++git ){
-                Shield<SEXP> tmp( obj->process_chunk(*git) ) ;
-                if( ! is<List>(tmp) || Rf_length(tmp) != 1){
-                    stop( "expecting a list of length 1" ) ;
-                }
-                res[i] = maybe_copy(VECTOR_ELT( tmp, 0)) ;
+        virtual bool handled(int i, const RObject& chunk ) {
+            if( is<List>(chunk) && Rf_length(chunk) == 1){
+                res[i] = maybe_copy(VECTOR_ELT(chunk, 0)) ;
+                return true ;
             }
-            return res ;
+            return false ;
+        }
+        virtual bool can_promote(const RObject& chunk ) {
+            return false ;    
+        }
+        virtual DelayedProcessor_Base<CLASS>* promote(int i, const RObject& chunk) {
+            return 0 ;    
+        }
+        virtual SEXP get() {
+            return res ;    
         }
 
     private:
-        int first_non_na ;
+        List res ;
 
         inline SEXP maybe_copy(SEXP x) const {
             return is_ShrinkableVector(x) ? Rf_duplicate(x) : x ;
@@ -124,21 +175,21 @@ namespace dplyr{
 
     } ;
 
-    template <typename CLASS, typename Data>
-    DelayedProcessor_Base<CLASS, Data>* get_delayed_processor(SEXP first_result, int i){
+    template <typename CLASS>
+    DelayedProcessor_Base<CLASS>* get_delayed_processor(int i, SEXP first_result, int ngroups){
         if( Rcpp::is<int>( first_result ) ){
-            return new DelayedProcessor<INTSXP, CLASS, Data>(i) ;
+            return new DelayedProcessor<INTSXP, CLASS>(i, first_result, ngroups) ;
         } else if( Rcpp::is<double>( first_result) ){
-            return new DelayedProcessor<REALSXP, CLASS, Data>(i) ;
+            return new DelayedProcessor<REALSXP, CLASS>(i, first_result, ngroups) ;
         } else if( Rcpp::is<Rcpp::String>( first_result) ){
-            return new DelayedProcessor<STRSXP, CLASS, Data>(i) ;
+            return new DelayedProcessor<STRSXP, CLASS>(i, first_result, ngroups) ;
         } else if( Rcpp::is<bool>( first_result) ){
-            return new DelayedProcessor<LGLSXP, CLASS, Data>(i) ;
+            return new DelayedProcessor<LGLSXP, CLASS>(i, first_result, ngroups) ;
         } else if( Rcpp::is<Rcpp::List>( first_result ) ){
             if( Rf_length(first_result) != 1 ) return 0 ;
-            return new DelayedProcessor<VECSXP, CLASS, Data>(i) ;
+            return new DelayedProcessor<VECSXP, CLASS>(i, first_result, ngroups) ;
         } else if( Rf_length(first_result) == 1 && TYPEOF(first_result) == CPLXSXP ){
-            return new DelayedProcessor<CPLXSXP, CLASS, Data>(i) ;
+            return new DelayedProcessor<CPLXSXP, CLASS>(i, first_result, ngroups) ;
         }
         return 0 ;
     }